Python HVC Camera Control
Got the python library to work on Python 3 so moved the camera logic into Python.

Still need to debug some of the camera connection stuff, check the logic on messy swapping interactions, and work out how to make the pi give enough power to run both cameras and the mic at the same time without dmesg showing over-current and the cameras dissapearing.